+++ django_project/core/adapter.py
@@ -1,5 +1,7 @@
 from account.hooks import AccountDefaultHookSet
-from django.conf import settings
+from django.core.mail import send_mail
+from django.template.loader import render_to_string
+from account.conf import settings
 
 
 class PinaxAccountHookset(AccountDefaultHookSet):
@@ -15,3 +17,9 @@ class PinaxAccountHookset(AccountDefaultHookSet):
             "host": settings.CONFERENCE_DOMAIN
         }
         super(PinaxAccountHookset,self).send_password_reset_email(to, newcontext)
+
+    def send_complete_profile_email(self,to,ctx):
+        subject = render_to_string("account/email/email_complete_profile_subject.txt",ctx)
+        subject = "".join(subject.splitlines())  # remove superfluous line breaks
+        message = render_to_string("account/email/email_complete_profile_message.txt", ctx)
+        send_mail(subject, message, settings.DEFAULT_FROM_EMAIL, to)
