+++ django_project/core/templates/account/email/email_complete_profile_subject.txt
@@ -1,0 +1,1 @@
+Complete Your Profile on FOSS4G 2017
